Slight generalization of the GH property fetching
The implementation of FetchProperty and FetchAllProperties for GitHub
was not generic enough as it required that the properties contain the
`name` key. For some properties, like pull request, the provider must
fetch by other keys, so let's just pass the properties to the wrapper
and let the wrapper error out if the get-by-properties don't contain the
necessary data for fetching the property.

Also moves the propertyOrigin struct to the generic fetcher.go source to
make it clearer that it's to be reused by other entities.