Save the refreshed PR properties, not the lookup ones
I mixed up the properties that we use to look up the PR and those that
we actually fetch from the provider. Let's save those that we use to
look up the PR to ensure that we have all the data available.

In this particular case, we were missing the SHA.