Fix two bugs in PR properties (#4346)
* Save the refreshed PR properties, not the lookup ones

I mixed up the properties that we use to look up the PR and those that
we actually fetch from the provider. Let's save those that we use to
look up the PR to ensure that we have all the data available.

In this particular case, we were missing the SHA.

* GitHub's PR number is not int64, but our protobuf is

Our protobuf defines the PR number as int64, but I didn't notice that
the GH API returns the PR number as plain int. It's easiest to upcast
the GH value to int64.