Create interface for registering entity through Provider
This is an interface that ensures that the provider does the necessary
steps to ensure Minder can track an entity. For GitHub and GitLab it
will be as a webhook for the repo or project. In DockerHub it'll be
a webhook for the image repository. Other providers might have other
steps.

The purpose is to define the interface, not implement it. That'll happen
in another PR.

Closes: #4325
Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>