Take properties into use in selectors and ruletype test.
Changes the ruletype test command to not use an EntityInfoWrapper as
intput, but just properties. An example of a repo entity with properties
is:
```
name: jakubtestorg/bad-python
is_private: false
is_fork: false
is_archived: false
upstream_id: "706108924"
github/repo_id:
  minder.internal.type: int64
    minder.internal.value: "706108924"
github/repo_name: bad-python
github/repo_owner: jakubtestorg
github/default_branch: master
foo: bar
```

Adds a lookup of an entity with properties to the evaluation engine so
that the resulting selectorEntity is augmented with properties and
passed to the selector engine. Changes the conversion interfaces to
selector entities to only use the EntityWithProperties model.

As a result, the user can use properties in selectors such as:
```
  - description: This should not apply copyleft repos
    entity: repository
    selector: repository.properties['github/license'] != 'GPL'
```
And in future allow us to extend selectors based on any other properties
we add.

Fixes: #4173