Implement general entity registration for the GitHub provider
This moves around the webhook handling code into the general entity
registration interface. Thus removing the need for the webhook manager.
Instead, each provider implements it's own registration, which _could_
be webhook-based.

The repo service was moved outside of the `github` package since it can
now work in a more general manner.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>