Take properties into use executor to fill selectors' properties (#4377)
* GetEntityByID only accepts an ID since it's unique

We can simplify the GetEntityByID call to not accept providerID since
the call searches on the globally-unique and indexed entity ID.

Related: #4173

* Add GetEntityWithPropertiesByID to our entity model package

Adds a convenience function to search an entity by ID with all its
properties. Also moves two utility functions form the property service
here.

Related: #4173

* Add ToProtoStruct to properties

This utility function will allow us to use properties both in selectors
that use the structupb.Struct and in the general entity protobuf message
that we added recently.

Fixes: #4173

* Add an option to NewProperties to skip checking the internal prefix

This is so unit tests can set int64 properties.

Related: #4173

* Take properties into use in selectors and ruletype test.

Changes the ruletype test command to not use an EntityInfoWrapper as
intput, but just properties. An example of a repo entity with properties
is:
```
name: jakubtestorg/bad-python
is_private: false
is_fork: false
is_archived: false
upstream_id: "706108924"
github/repo_id:
  minder.internal.type: int64
    minder.internal.value: "706108924"
github/repo_name: bad-python
github/repo_owner: jakubtestorg
github/default_branch: master
foo: bar
```

Adds a lookup of an entity with properties to the evaluation engine so
that the resulting selectorEntity is augmented with properties and
passed to the selector engine. Changes the conversion interfaces to
selector entities to only use the EntityWithProperties model.

As a result, the user can use properties in selectors such as:
```
  - description: This should not apply copyleft repos
    entity: repository
    selector: repository.properties['github/license'] != 'GPL'
```
And in future allow us to extend selectors based on any other properties
we add.

Fixes: #4173

* Move GetEntityWithPropertiesByID from models to service