Derive the name property from entity name for entities that have not migrated fully
We have entities in the database that have an entity instance, but no
corresponding property because they have not been migrated properly. In
that case, we can't use properties to refresh the entity because the
properties are not there. The entity itself has a name, though, so let's
use it to backfill the name property so that the properties object can
always be used to refresh the rest.