Fix entity refresh for entities that have not been migrated (#4378)
* Add a setter for properties

Adds a utility function to set a property so that we don't have to
construct a map to construct properties to merge properties.

* Derive the name property from entity name for entities that have not migrated fully

We have entities in the database that have an entity instance, but no
corresponding property because they have not been migrated properly. In
that case, we can't use properties to refresh the entity because the
properties are not there. The entity itself has a name, though, so let's
use it to backfill the name property so that the properties object can
always be used to refresh the rest.

* Improve error message

This would actually tell us what entity failed to refresh