Implement caching of providers on the provider manager (#4400)
This generalized the expiration cache that was first introduced for REST
clients and introduced it as a general caching mechanism.

Subsequently, this was used to implement a 1 min cache for provider
fetching. This optimizes for cases where the provider is fetched
frequently for a single operation. Now there's no need to go multiple
times to the database.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>