Add `properties` to Repositories protobuf message
This adds properties which will allow us to output the entity properties
as we're gathering them. Even though we're hoping to get rid of the
`Repositories` message entirely, this is handy to have while we make the
transition.

Note that there is nothing outputting this nor using this at the moment.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>