Call an explicit commit in ListEvaluationHistory handler
We refresh the properties in the historical evaluation endpoint (which
btw should just be temporary to convert from the old tables to the new
general tables) and missed the fact that the handler starts a
transaction and then always rolls back - that was probably a way to keep
consistence.

This means that the transactions that refresh properties are rolled back
and on next call fetched again. This hurts performance quite a bit on
repeated listHistory calls.

Fixes: #4393