Don't fail if lookupProperties include upstream ID but the entity does not
We recently enhanced the property service to include lookup by ID in the
cache, which is great, but we didn't handle the case where the lookup
properties included the upstream ID but the entity properties in the
database did not. In that case we should fall back to looking the
property by name.