Fix retrieving properties by upstream ID and name (#4432)
* Don't fail if lookupProperties include upstream ID but the entity does not

We recently enhanced the property service to include lookup by ID in the
cache, which is great, but we didn't handle the case where the lookup
properties included the upstream ID but the entity properties in the
database did not. In that case we should fall back to looking the
property by name.

* Don't return provifv1.ErrEntityNotFound from the property service

It was irritating to have to handle two return codes. The caller really
doesn't care why was the entity not found, just that it was not found.
Let's conver the provider NotFound into service Not found so that the
callers have one less error to handle.