The Go GitHub API doesn't seem to escape user package names, only org package names
For some reason, the Go github API doesn't escape package names when
listing user packages:
https://github.com/google/go-github/blob/662da6f8e9f32b7da649ad0bfac19948e5acdd85/github/users_packages.go#L58
but does so when listing org packages:
https://github.com/google/go-github/blob/662da6f8e9f32b7da649ad0bfac19948e5acdd85/github/orgs_packages.go#L48

This tripped us up when trying to refresh properties of a package that
belonged to a user and had e.g. a slash in the name.

It seems that evaluating these packages did not work since we upgraded to
github-go v60 as we had the same bug in the ingester and the properties
changes just made the bug more prominent in the sense that it would manifest
already in the handlers.

Fixes: #4444