Add "options" to properties service calls
Instead of passing the db store and other optional values as
explicit paramters, this introduces options to the properties service.
This allows us to start building queries in a more dynamic way.

The querier is passed always now as an option.

Disregarding stale data was introduced as an option, and it was also
used in the places it's needed, that is, in GET calls.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>