Add general webhook registration mechanism (#4476)
This adds per provider class paths so each provider class will bring
it's own webhook handling. GitHub is a slightly special case... so we
handle it separately for now.

Note that this will require folks changing their configuration to remove
`github` as a suffix on deployments. If nothing is changed, this should
be a no-op as it does not re-append `github` in the path.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>