Include the name of the package for which we couldn't find a fix
When replying to a single line, github always prepends the context "before"
the line unless replying with a suggestion. This gets confusing when replying
to a line deeper in a diff especially the last line as we can't pinpoint a
single line to reply to.

To improve the UX a little bit, we include the package name with the
package we are replying to so that the user can tie the comment better
to the actual line.

Pair-Programmed-With: Michelangelo Mori <mmori@stacklok.com>