Add `ReregisterEntity` to provider interface (#4541)
This adds an interface that's meant to re-register an entity. This is
useful when updating a certain part of the entity registration. The only
use-case we have right now is secret rotation. So this is what was
effectively implemented.

This merely implemented the functions for the GitLab and GitHub
providers. The webhook update sub-command will be updated to take it
into use in a subsequent patch.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>