Clean up test providers and move `EntityToProtoMessage` as top-level interface function
We were doing a lot of repetition in our test providers, so this
implemented a `noop` provider that fills in the signature as needed.

This also moved the `EntityToProtoMessage` function to be a top-level
provider function in the interface. While this is temporary, it's useful
to know this needs to be defined while we move everything to the general
protobuf interface.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>