GitHub: don't try to use OAuth token for github-app provider
This separates the providers a little further by always assuming that if
the github app provider class is used, you should only check for the
github app credentials (which require an installation ID).

This will thus reduce the number of database calls we do when
instantiating the provider, since Minder will now do the right thing.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>