gitlab: Handle token refresh
Gitlab uses shorter lived tokens than github, this adds the logic to
handle a token refresh.

At the moment this is done within the gitlab provider manager. While
this works for now, we might want to determine whether we make this a
more general pattern, or keep handling this within each provider.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>