Use generic message forwarding to initialize the profile instead of entity-specific database searches
Because we were bound to entity-specific protobuf types and didn't have
a good way of iterating over all entities, we used to enumerate all
repos and then all artifacts just to marshall them and send off to the
evaluator. With generic entities we can simply enumerate all entities
and send them off for refresh and forwarding to the evaluator.

Fixes: #4621