Do not forward messages for archived and/or private repos
Since the github webhook handler won't have access to the properties and
the properties will be refreshed and read later on, we'd lose the
ability to not evaluate messages from private or archived repositories.

We do check for the private and archived flags at repo registration, but
since a repo can go private or be archived later, this is not enough.

Let's check for these properties after retrieving the entity and before
forwarding it on.

Fixes: #4666