Remove reconciling of artifacts of repo reconciliation
The GitHub API doesn't allow us to list packages for repos, only for
organizations. We tried to work around that by listing packages for the
org and then filtering by the repository attribute, but this does not scale well.

Moreover, we agreed that going forward we'd like to register and enroll
packages in a similar way that we do register and enroll repositories.

Fixes: #4618