Use the handler forwarding to evaluate repository/meta messages
Instead of fetching the repo directly in the handler and forwarding
to a handler, create a NewEntityRefreshAndDoMessage message and offload
the entity processing to minder-core.

This removes the last piece of the github webhook that was directly
touching the database with the exception of looking up the github app
installation.

Fixes: #4703