Don't look up the same entity multiple times in ListEvaluationHistory
We were looking up the same entity over and over again in
ListEvaluationHistory. Let's look up each entity just once, then cache
the properties.

This doesn't really increase the memory usage of the handler as we'd be
keeping the entity around anyway. With this change, we just keep a map
of UUID-to-pointer around for faster lookups.

Related: #4690