Fix race in github webhook handler tests.
Recent changes introduced a couple additional cases in which we
forward events from the webhook handler to Minder core, causing helper
utility `PassthroughQueue` to get stuck sending on a channel that was
not consumed anymore.

This change should fix the issue by
* consuming the new messages (and running assertions over them)
* ensuring no additional messages are there to be consumed

Additionally, in case tests fail to consume all produced messages, the
test now panics if ran with `-race`.