Implement an entity cache for the properties service (#4729)
This adds a wrapper that is able to cache entities for cases that need
to iterate them. This is handy for one-off calls like getting the
history of executions or showing reports. This way, we don't need to
query the database multiple times.

This cache is persistent, so it's only relevant for one execution.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>