Move the GitHub webhook handler into the github provider
This commit moves over the webhook handler and its tests from
controlplane to the github provider. There should be no functional
changes except the github provider class manager now accepts metrics and
an eventer as parameters to be able to emit watermill events.

The github app and marketplace handlers are a special-case which we
might refactor later.

One test to check the content size middleware was moved off the github
webhook handler test and into controlplane tests.

Fixes: #4328