Introduce `ResourceMeta` and general resource parsing
This introduces the concept of resource metadata, which includes three
things:

* `type`: the type of resource
* `version`: The version of a resource
* `name`: The name of a resource

Parsing these three is mandatory for any resource in Minder's domain
model.

This also introduces a general parser which replaces the
rule-type-specific parser.

Finally, the rule type `create`/`apply` commands were refactored to use
this new parser, and will skip files that are effectively not rule
types. Thus allowing the aforementioned sub commands to run on file
directories with objects that aren't necessarily rule types.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>