Make `minder ruletype` `apply`/`create` smarter
This changes the behavior of these commands...

If we only give one file or standard input, it will actually parse
and fail if it's not a minder resource (you probably REALLY wanted
to apply that file).

If it's a directory, it'll try to be smart and apply as much as it can.

It all depends on if the file was expanded or not.

Consequently, this fixes the issue we used to have with test files in
the `minder-rules-and-profiles` repo.

Co-Authored-By: Michelangelo Mori <mmori@stacklok.com>
Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>