Expose a helper to run a JQ expression on YAML input through our Rego library (#4793)
* Add JQEvalBoolExpression

This adds a helper utility that takes a JQ expression that is to return
a boolean and runs the expression againts the provided object. The
expression must return a boolean.

* Expose JQEvalBoolExpression through Rego

Exposes the JQEvalBoolExpression function added in the previous patch
through our Rego library. This is to make it usable by rule types.

* Split the previous interfaces into two - one that loads the YAML and one that runs the JQ query