Add YQ-evaluating remediator
In order to be able to change YAML files such as github workflows safely and
with minimal amount of changes, we need to add a new remediation
function in addition to the put-a-content-somewhere and call-frizbee
ones we have now.

I chose to add one based on `libyq` which does a decent job at retaining
comments and general YAML structure.

Fixes: #4815