Extend the PR evaluator with generic params
The PR evaluator used to have typed params for every function we'd add -
there's one for frizbee, one for the pull_request_content function etc.
This is not great, as we tie the functions to our protobuf API and every
new function requires a client release to be done or else the clients
can't even add ruletypes with the new functions.

Let's just use a generic `structpb.Struct` going forward. This time it
is still a change that needs client support, but going forward we'll
just have to change the server code to add a new function.