Exclude profiles with labels from history. (#4910)
Filter by labels when listing history.

History list showed evaluation results even for profiles having
labels. This is different from the behaviour of Profile list and
status, which by default only showed results for profiles without
labels.

This change makes the behaviour of History list closer to the one
Profile list and status by adding a new filter to the
`ListEvaluationHistory` endpoint. Both inclusion and exclusion are
supported, and the additional `*` flag is allowed specifically for
inclusion.

Fixes stacklok/minder-stories#102