Implement data source interfaces & registration in rego engine
This implements the basic interfaces that data sources must fulfil in
order to be called and used within Minder. This is:

* Listing the functions they provide
* Validating updates

Each function itself must provide a unique key for the engine to use, as
well as argument validation and the data source call itself.

Finally, this implements an engine option that allows data sources to be
dynamically registered into the rego engine.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>