Add data sources to ruletype API (#5019)
* Add data sources to ruletype API

This adds the ability for the eval engine to reference data sources to
be instantiated. this allows us to explicitly declare what data sources
a rule type needs which will in turn be used to keep track of
dependencies.

Only declared ruletypes will be registered in the per-rule data source
registry.

Data sources are declared in ruletypes by name. The API was built in a
way that we can eventually add other reference means other than names;
but names will do for now.

Note that name validation was also changed to reflect that we only
accept lowercase characters for data source names.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>

* Gate data source references in rule types with feature flag

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>

---------

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>