Validate Data Sources when creating rules.
Rules can now specify a list of data sources names that they need for
their evaluation. This change ensures that data sources required by
the rule are available when creating the rule.

A similar change is being added in the code path implementing data
source deletion ensuring that a data source is not deleted if there
exist rules depending on it.

Fixes #5049