Fix on delete cascade in `rule_type_data_sources`.
This migration recreates the foreign key constraint to delete rows
from `rule_type_data_sources` when a record is deleted from
`rule_type`.

Note that it is safe to just drop and recreate the constraint as the
previous version prevented the deletion of rule types if they
referenced a data source, so it was not possible to have dangling
records.