data sources: Validate individual function updates
This hooks in the update validation that's enforced by the data sources
interfaces. This means, we validate that if you do an update to a data
source the update is backwards compatible.

We do this by checking that you're not deleting functions via an update
and that any functions you modify pass our schema update validation.

In the future, we could allow folks to delete functions via a `force`
flag in the API.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>