Simplify propagation of action settings
Currently, action settings are passed around via a complex dance of
getter/setter functions that span multiple places in the executor.

However, this is not necessary, as we've already read these settings at
the beginning of execution, and can just set them in place where we need
them.

This simplifies this logic and leverages the aggregate profile structure
to set this. It also enforces the setting as part of the action driver
initiatlization.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>