Allow REST data sources to form bodies from input (#5130)
This adds a new option to the REST data source driver to form a body
from input. This will be marshaled as JSON or interpreted as a string,
and will be used as the message body.

This allows us to dynamically form the body in order to better support
POST-based implementations like OSV.

the new field is called `body_from_field`. The field *must* exist in the
input parameters and must be marked as required or have an appropriate
default. The field must also be a top-level parameter.

Here's an example of a data source that uses it:

```yaml

---
version: v1
type: data-source
name: osv
context: {}
rest:
  def:
    query:
      # Supports templating based on RFC6570
      endpoint: 'https://api.osv.dev/v1/query'
      parse: json
      method: POST
      body_obj_from_field: query
      input_schema:
        type: object
        properties:
          query:
            type: object
            properties:
              version:
                type: string
              package:
                type: object
                properties:
                  ecosystem:
                    type: string
                    description: The ecosystem the dependency belongs to
                  name:
                    type: string
                    description: The name of the dependency
        required:
          - query
```

Note in this case we do validation of all parameters that can be passed
to the body.

usage within a rego evaluation policy would look as follows:

```
...

          reqparams := {
              "query": {
                "version": version,
                "package": {
                  "name": name,
                  "ecosystem": "PyPI"
                }
              }
            }

          out := minder.datasource.osv.query(reqparams)
	  ...
```

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>