Promote pull request properties
The source and target PR information, as well as the upstream URL have
been promoted to top level properties.

The GitLab provider now fills these in by default and uses them to
populate the PR protobuf.

The GitHub provider fills in the properties in, but currently does not
leverage them; this will be done after some rollout time.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>