Add REGO debugger to Mindev.
This change adds the possibility to start evaluate a REGO-based rule
type in a debugger.

The debugger allows setting breakpoints, stepping, printing source,
and a few other simple utilities.

The debugger is currently very, very, VERY rough around the edges and
could use some love, especially in the reception of events from the
debuggee, which is done inline and not asynchronously.