Enforce defaults in `deps` ingest pull request configuration
If the user would forget to set the parameters, the deps ingest would
fail saying that the filter is invalid. This is not ideal for the user
since we should just set reasonable defaults.

This fixes that!

We now default to `new_and_updated`  if no filter configuration is set.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>