Refactor: Move Properties to pkg as Public API
Properties are leveraged by many pieces of Minder, including providers, and should be considered a public API instead of being in our internal package.
The idea is to move properties to `pkg`.
However, the underlying service package will remain internal, as it is a minder-server implementation detail.

This commit fixes the affected functions to refer to the public API to access properties

Signed-off-by: Kugamoorthy Gajananan <gajananan@gmail.com>