Refactor: Move Properties to pkg as Public API
Properties are leveraged by many pieces of Minder, including providers, and should be considered a public API instead of being in our internal package.
The idea is to move properties to `pkg`.
However, the underlying service package will remain internal, as it is a minder-server implementation detail.

This commit includes changes after running
1.  `make bootstrap` again to resolve conflicts in the following files
- internal/proto/internal.pb.go
- pkg/api/protobuf/go/minder/v1/minder.pb.go

2. `make gen`
Signed-off-by: Kugamoorthy Gajananan <gajananan@gmail.com>